+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 - Changed default filter bindings changed to make projectile not a hard dependency
 - Added `frog-jump-buffer-include-current-buffer` option
 - Allow for overriding `posframe` related options
+
+*** 0.1.2
+
+- Add support for filtering by `recentf` and opening closed buffers
+- Change default filter bindings again to make room for `recentf`